% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myclt.R
\name{myclt}
\alias{myclt}
\title{myclt}
\usage{
myclt(n, iter)
}
\arguments{
\item{int}{iter the number of samples}
}
\value{
hist(sm/n) a histogram of the sample means and sm/n the vector of sample means
}
\description{
myclt
}
